Initial release is under review on Google Play Store

+++ Muvlex/app/src/main/java/com/martiandeveloper/muvlex/model/Movie.kt
@@ -31,7 +31,7 @@ data class Movie(
     @Json(name = "video")
     val video: Boolean?,
     @Json(name = "vote_average")
-    val voteAverage: Float?,
+    val voteAverage: Double?,
     @Json(name = "vote_count")
     val voteCount: Int?
 ) : Parcelable
